added:
NEWS
CITATION
link to help manaual in README
url link for bug reprot in DESCRIPTION

Dason the citation and news are just to get them in there.  Please alter them anyway you want.  I'm not sure how tomake two institutions appear for the citation so we may want to drop that.