Updated `p_vignette` to utilize `tools::browseVignettes`.  In the process I renamed `print.p_data` to `print.wide_table` to allow any pacman data.frame/table/matrix that is expected to be wide and character vector to utilize this print class.
One point I'm not happy about but do not know how to fix.  I like the default (no packages supplied) to give all the packages.  I made this so multiple packages could be passed htrough \ldots.  The problem is that for:

p_vignette(interactive = FALSE)

This causes the ldots to read the FALSE and thus `packs == "FALSE"` as a character vector.  I used some logical if(){} to fix this:

(length(pack) == 1 && pack %in% c("TRUE", "FALSE"))

But it seems there would be a better approach that I'm not thinking of.

close #33