Added a get option to `p_citation` as to whether or not to copy the tex or standard to the clipboard.
We may want to think of other arguments that would benefit from setting an option and defaulting to behaviors most our users (or more importantly ourselves :) would want.  These would be arguments that are part of a user's standard work flow.  If you use BibTex today you're likely to throughout and vise versa.