re-wrote p_delete to use a single version `p_delete_single` per Dason's request https://github.com/trinker/pacman/issues/11#issuecomment-6335656
I used a meta data approach in `p_delete_single` that invisibly returns:

data.frame(package, can_unload, type)

This gives info on package name (character); can it be deleted (logical); and if can not be deleted the type (base or not a package).

I checked that p_delete works and it appears to.