Removed the print method and class for p_delete.  it makes it unnecessarily complicated.  I truned the print.p_delete function into an internal function p_delete_warning that reduces the code inside of p_delete.
I also added the is.global function to pacman.  It breaks the p_xxx frame work but in the historical context of R with naming conventions, is more sensible and desscriptive.  This function is intended to be used more as a default argument to detect global environment.  That way if it is used inside a function, its default is not to print (or what ever action it is the default argument to).