Added: p_iscran to the p_cran rdname for logical check of cran-ness
Reformatted p_search_library to seach by `begins.with` and `contains`.  I removed the ability of the function to take non character strings for search terms.  THis is consistent with the rest of the package (we only allow this for package names).  This function is lighter weight, safer, and more compact.

Made some minor documentation changes to various pacman functions.

Finished first draft of intro vignette.